@@ -1,4 +1,6 @@
-﻿using Microsoft.Extensions.DependencyInjection;
+﻿using PublicTransportation.Domain.Interfaces.Repositories;
+using PublicTransportation.Infra.Repository;
+using PublicTransportation.Application.UseCases.Lines;
 
 namespace PublicTransportation.Api.Configuration
 {
@@ -6,6 +8,11 @@ namespace PublicTransportation.Api.Configuration
     {
         public static void RegisterServices(this IServiceCollection services)
         {
+            //repositories
+            services.AddScoped<ILineRepository, LineRepository>();
+
+            //services
+            services.AddScoped<LineServices>();
         }
     }
 }
